+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 - Documentation on valid client types (#2015)
 - Added `compiler.json` option and `--json` compile flag to export structured compile data such as occurred errors (#1206)
 - Added troubleshooting documentation (#1211)
+- Documentation on compiler API and JSON (#2060)
+- Added further documentation to inmanta-cli command (#2057)
+
 
 # v 2020.2 (2020-04-24) Changes in this release: